refactor: reorganize codebase into application_commands architecture
Restructure Discord command handlers following discord.py cogs pattern:
- Move handlers/register.ts → application_commands/register/handler.ts
- Move shared utilities to application_commands/shared/
- Update imports across codebase to match new structure
- Rename register.ts → utils/command-registration.ts

Establishes scalable architecture for adding new Discord commands
while avoiding naming conflicts between utilities and command handlers.
