+++ package.json
@@ -28,9 +28,9 @@
     "typecheck": "tsc --noEmit --strict",
     "type-coverage": "type-coverage --at-least 95 --strict --ignore-files '**/__tests__/**'",
     "build": "tsc && wrangler deploy --dry-run",
-    "register": "node --env-file=.dev.vars --import tsx src/register.ts",
-    "register:dev": "DISCORD_ENV=development node --env-file=.dev.vars --import tsx src/register.ts && echo 'Commands registered for development'",
-    "register:prod": "DISCORD_ENV=production node --env-file=.dev.vars --import tsx src/register.ts && echo 'Commands registered for production'",
+    "register": "node --env-file=.dev.vars --import tsx src/utils/command-registration.ts",
+    "register:dev": "DISCORD_ENV=development node --env-file=.dev.vars --import tsx src/utils/command-registration.ts && echo 'Commands registered for development'",
+    "register:prod": "DISCORD_ENV=production node --env-file=.dev.vars --import tsx src/utils/command-registration.ts && echo 'Commands registered for production'",
     "setup:dev": "cp .dev.vars.example .dev.vars && echo 'Please edit .dev.vars with your Discord credentials'",
     "setup:tunnel": "./scripts/setup-tunnel.sh",
     "save:tunnel": "node scripts/capture-tunnel.js",
